solve int and unsigned int

+++ src/rasterizer/pipeline.cpp
@@ -630,7 +630,7 @@ void Pipeline<p, P, flags>::rasterize_triangle(
     std::array<float, 5> q_a_attributes = {0,0,0,0,0};
     std::array<float, 5> q_a_up_attributes = {0,0,0,0,0};
     std::array<float, 5> q_a_right_attributes = {0,0,0,0,0};
-    for (int i = 0; i < sizeof(va.attributes)/sizeof(va.attributes[0]); ++i) {
+    for (int i = 0; i < (int)(sizeof(va.attributes)/sizeof(va.attributes[0])); ++i) {
         q_a_attributes[i] = q_a_barycentric.x * va.attributes[i] + q_a_barycentric.y * vb.attributes[i] + q_a_barycentric.z * vc.attributes[i];
         q_a_up_attributes[i] = q_a_up_barycentric.x * va.attributes[i] + q_a_up_barycentric.y * vb.attributes[i] + q_a_up_barycentric.z * vc.attributes[i];
         q_a_right_attributes[i] = q_a_right_barycentric.x * va.attributes[i] + q_a_right_barycentric.y * vb.attributes[i] + q_a_right_barycentric.z * vc.attributes[i];
